+++ src/main/kotlin/com/modernflow/keyvaluestore/proxy/dtos/KeyValueStoreDto.kt
@@ -1,4 +1,4 @@
-package com.example.keyvaluestore.dtos
+package com.modernflow.keyvaluestore.proxy.dtos
 
 data class KeyValueStoreDto(
     val key: String,
