+++ src/main/kotlin/com/modernflow/keyvaluestore/proxy/dtos/StoreUpsertRequestDto.kt
@@ -1,4 +1,4 @@
-package com.example.keyvaluestore.dtos
+package com.modernflow.keyvaluestore.proxy.dtos
 
 data class StoreUpsertRequestDto(
     val value: String,
