+++ src/main/kotlin/com/modernflow/keyvaluestore/proxy/services/StoreUpsertService.kt
@@ -1,7 +1,7 @@
-package com.example.keyvaluestore.services
+package com.modernflow.keyvaluestore.proxy.services
 
-import com.example.keyvaluestore.dtos.KeyValueStoreDto
-import com.example.keyvaluestore.store.ConsistenceHashMap
+import com.modernflow.keyvaluestore.proxy.dtos.KeyValueStoreDto
+import com.modernflow.keyvaluestore.proxy.store.ConsistenceHashMap
 import org.springframework.stereotype.Service
 
 @Service
