+++ src/main/kotlin/com/modernflow/keyvaluestore/proxy/store/ConsistenceHashMap.kt
@@ -1,6 +1,6 @@
-package com.example.keyvaluestore.store
+package com.modernflow.keyvaluestore.proxy.store
 
-import com.example.keyvaluestore.hash.Hash
+import com.modernflow.keyvaluestore.proxy.hash.Hash
 import org.springframework.stereotype.Component
 import java.util.*
