+++ src/test/kotlin/com/modernflow/keyvaluestore/proxy/KeyValueStoreApplicationTests.kt
@@ -1,4 +1,4 @@
-package com.example.keyvaluestore
+package com.modernflow.keyvaluestore.proxy
 
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.context.SpringBootTest
